packages/topCitiesStats/requirements.txt
SQLAlchemy
PyMySQL
python-dotenv
functions_framework